+++ quickstart/tests.py
@@ -1,6 +1,6 @@
 from django.test import SimpleTestCase
 from telegram_framework import get_bot, actions, messages
-from telegram_framework.chat import Chat, add_bot, get_last_message
+from telegram_framework import chats
 from telegram_framework.links import add_links
 from quickstart.bot import bot_links
 
@@ -9,17 +9,17 @@ class TestCommands(SimpleTestCase):
 
     def setUp(self):
         # Создайте чат для тестовых сообщений
-        chat = Chat()
+        chat = chats.Chat()
         # Создайте тестового пользователя
         self.client = get_bot('client')
         # Добавьте его в чат
-        chat = add_bot(chat, self.client)
+        chat = chats.add_bot(chat, self.client)
         # Создайте пользователя - бота
         bot = get_bot('bot')
         # Свяжите бота с его обработчиками
         bot = add_links(bot, bot_links)
         # Добавьте его в чат
-        self.chat = add_bot(chat, bot)
+        self.chat = chats.add_bot(chat, bot)
         # В чате пока нет сообщений
         self.assertEqual(0, len(self.chat.messages))
 
@@ -37,7 +37,7 @@ class TestCommands(SimpleTestCase):
         # Поэтому в чате будет 2 сообщения
         self.assertEqual(2, len(chat.messages))
         # Получите последнее сообщение для проверки
-        last_message = get_last_message(chat)
+        last_message = chats.get_last_message(chat)
         expected_text = 'Приветствую тебя. Я Quickstart Telegram Bot'
         # Оно должно содержать приветствие
         self.assertEqual(expected_text, last_message.text)
@@ -56,7 +56,7 @@ class TestCommands(SimpleTestCase):
         # Поэтому в чате будет 2 сообщения
         self.assertEqual(2, len(chat.messages))
         # Получаем последнее сообщение
-        last_message = get_last_message(chat)
+        last_message = chats.get_last_message(chat)
         expected_text = 'Тебе отвечает Bot'
         # Оно должно содержать ответ бота
         self.assertEqual(expected_text, last_message.text)
