+++ telegram_framework/py_telegram_bot_api/actions.py
@@ -1,4 +1,4 @@
-from telegram_framework.chat import add_message, Chat
+from telegram_framework import chats
 from telegram_framework.messages import Message, Reply
 
 
@@ -8,16 +8,16 @@ def send_reply(reply: Reply):
     chat = message.chat
     bot = reply.sender
     bot.reply_to(message, text)
-    chat = Chat(id=chat.id)
-    return add_message(chat, reply)
+    chat = chats.Chat(id=chat.id)
+    return chats.add_message(chat, reply)
 
 
-def send_message(chat: Chat, message: Message):
+def send_message(chat: chats.Chat, message: Message):
     bot = message.sender
     text = message.text
     bot.send_message(
         chat.id,
         text,
     )
-    chat = Chat(id=chat.id)
-    return add_message(chat, message)
+    chat = chats.Chat(id=chat.id)
+    return chats.add_message(chat, message)
