+++ telegram_framework/tests/test_chat.py
@@ -3,7 +3,7 @@ Tests
 """
 from dataclasses import dataclass
 from django.test import SimpleTestCase
-from telegram_framework.chat import Chat, add_bot, get_last_message, add_message
+from telegram_framework import chats
 from telegram_framework.messages import Message
 
 
@@ -23,7 +23,7 @@ class TestChat(SimpleTestCase):
         """
         setUp test data
         """
-        self.chat = Chat()
+        self.chat = chats.Chat()
 
 
     def test_add_bot(self):
@@ -32,7 +32,7 @@ class TestChat(SimpleTestCase):
         """
         bot = EmptyBot()
         self.assertEqual(0, len(self.chat.bots))
-        chat = add_bot(self.chat, bot)
+        chat = chats.add_bot(self.chat, bot)
         self.assertEqual(1, len(chat.bots))
         chat_bot = chat.bots[0]
         self.assertEqual(bot, chat_bot)
@@ -43,7 +43,7 @@ class TestChat(SimpleTestCase):
         """
         self.assertEqual(0, len(self.chat.messages))
         message = Message(text='some text', sender='some sender')
-        chat = add_message(self.chat, message)
+        chat = chats.add_message(self.chat, message)
         self.assertEqual(1, len(chat.messages))
 
     def test_add_message_in_chat(self):
@@ -52,8 +52,8 @@ class TestChat(SimpleTestCase):
         """
         message = Message(text='some text', sender='some sender')
         self.assertIsNone(message.chat)
-        chat = add_message(self.chat, message)
-        chat_message = get_last_message(chat)
+        chat = chats.add_message(self.chat, message)
+        chat_message = chats.get_last_message(chat)
         self.assertEqual(message, chat_message)
         self.assertEqual(chat, chat_message.chat)
 
@@ -61,7 +61,7 @@ class TestChat(SimpleTestCase):
         """
         Test get_last_message: success: from empty chat
         """
-        message = get_last_message(self.chat)
+        message = chats.get_last_message(self.chat)
         self.assertIsNone(message)
 
     def test_get_last_message(self):
@@ -69,9 +69,9 @@ class TestChat(SimpleTestCase):
         Test get_last_message: success
         """
         message = Message(text='some text', sender='some sender')
-        chat = add_message(self.chat, message)
+        chat = chats.add_message(self.chat, message)
         self.assertEqual(1, len(chat.messages))
-        last_message = get_last_message(chat)
+        last_message = chats.get_last_message(chat)
         self.assertEqual(message, last_message)
         self.assertEqual(last_message.chat, chat)
         self.assertEqual(len(last_message.chat.messages), len(chat.messages))
