+++ telegram_framework/tests/test_dummy/test_actions.py
@@ -1,6 +1,6 @@
 from django.test import SimpleTestCase
 from telegram_framework.dummy.actions import send_message, send_reply
-from telegram_framework.chat import Chat, get_last_message, add_message
+from telegram_framework import chats
 from telegram_framework.messages import Message, Reply, create_reply
 
 
@@ -10,26 +10,26 @@ class TestActions(SimpleTestCase):
         """
         Test send_message: success
         """
-        chat = Chat()
+        chat = chats.Chat()
         self.assertEqual(0, len(chat.messages))
         message = Message('new message', sender='some sender')
         chat = send_message(chat, message)
         self.assertEqual(1, len(chat.messages))
-        last_message = get_last_message(chat)
+        last_message = chats.get_last_message(chat)
         self.assertEqual(message, last_message)
 
     def test_send_reply(self):
         """
         Test send_reply: failed: message has no chat
         """
-        chat = Chat()
+        chat = chats.Chat()
         self.assertEqual(0, len(chat.messages))
         message = Message('new message', sender='some sender')
-        chat = add_message(chat, message)
-        last_message = get_last_message(chat)
+        chat = chats.add_message(chat, message)
+        last_message = chats.get_last_message(chat)
         reply = create_reply(last_message, 'reply', sender='other sender')
         chat = send_reply(reply)
-        last_reply = get_last_message(chat)
+        last_reply = chats.get_last_message(chat)
         expected_reply = Reply(
             'reply',
             'other sender',
