+++ telegram_framework/tests/test_messages/test_reply.py
@@ -2,7 +2,7 @@ from django.test import SimpleTestCase
 from telegram_framework.messages.errors import MessageNotInChatError
 from telegram_framework.messages.reply import Reply, create_reply
 from telegram_framework.messages import Message
-from telegram_framework.chat import Chat, add_message, get_last_message
+from telegram_framework import chats
 
 class TestReply(SimpleTestCase):
 
@@ -18,10 +18,10 @@ class TestReply(SimpleTestCase):
         """
         Test create_reply: success
         """
-        chat = Chat()
+        chat = chats.Chat()
         message = Message('new message', sender='some sender')
-        chat = add_message(chat, message)
-        last_message = get_last_message(chat)
+        chat = chats.add_message(chat, message)
+        last_message = chats.get_last_message(chat)
         reply = create_reply(last_message, 'reply', sender='other sender')
         expected_reply = Reply(
             'reply',
