+++ telegram_framework/tests/test_py_telegram_bot_api/test_actions.py
@@ -1,7 +1,7 @@
 # pylint: disable=duplicate-code
 from django.test import SimpleTestCase
 from telegram_framework.py_telegram_bot_api.actions import send_message, send_reply
-from telegram_framework.chat import Chat, get_last_message, add_message
+from telegram_framework import chats
 from telegram_framework.messages import Message, Reply, create_reply
 
 
@@ -23,26 +23,26 @@ class TestActions(SimpleTestCase):
         """
         Test send_message: success
         """
-        chat = Chat()
+        chat = chats.Chat()
         self.assertEqual(0, len(chat.messages))
         message = Message('new message', sender=self.bot)
         chat = send_message(chat, message)
         self.assertEqual(1, len(chat.messages))
-        last_message = get_last_message(chat)
+        last_message = chats.get_last_message(chat)
         self.assertEqual(message, last_message)
 
     def test_send_reply(self):
         """
         Test send_reply: failed: message has no chat
         """
-        chat = Chat()
+        chat = chats.Chat()
         self.assertEqual(0, len(chat.messages))
         message = Message('new message', sender=self.bot)
-        chat = add_message(chat, message)
-        last_message = get_last_message(chat)
+        chat = chats.add_message(chat, message)
+        last_message = chats.get_last_message(chat)
         reply = create_reply(last_message, 'reply', sender=self.bot)
         chat = send_reply(reply)
-        last_reply = get_last_message(chat)
+        last_reply = chats.get_last_message(chat)
         expected_reply = Reply(
             'reply',
             self.bot,
